core/fi_peer: Move peer provider defines into new header
The peer provider APIs target provider implementations, not
applications.  Move the definitions out of fi_ext.h, which is
used by apps, and into include/rdma/providers/fi_peer.h.  This
better indicates that the APIs are intended for providers, not
applications.
