+++ Makefile.am
@@ -230,8 +230,9 @@ rdmainclude_HEADERS += \
 	$(top_srcdir)/include/rdma/fi_tagged.h \
 	$(top_srcdir)/include/rdma/fi_trigger.h
 providersinclude_HEADERS += \
-	$(top_srcdir)/include/rdma/providers/fi_prov.h \
-	$(top_srcdir)/include/rdma/providers/fi_log.h
+	$(top_srcdir)/include/rdma/providers/fi_log.h \
+	$(top_srcdir)/include/rdma/providers/fi_peer.h \
+	$(top_srcdir)/include/rdma/providers/fi_prov.h
 
 if HAVE_DIRECT
 nodist_rdmainclude_HEADERS = \
