+++ libfabric.vcxproj
@@ -941,6 +941,7 @@
     <ClInclude Include="include\rdma\fi_trigger.h" />
     <ClInclude Include="include\rdma\fi_collective.h" />
     <ClInclude Include="include\rdma\providers\fi_log.h" />
+    <ClInclude Include="include\rdma\providers\fi_peer.h" />
     <ClInclude Include="include\rdma\providers\fi_prov.h" />
     <ClInclude Include="include\uthash.h" />
     <ClInclude Include="include\windows\arpa\inet.h" />
