+++ prov/shm/src/smr.h
@@ -51,7 +51,7 @@
 #include <rdma/fi_tagged.h>
 #include <rdma/fi_trigger.h>
 #include <rdma/providers/fi_prov.h>
-#include <rdma/fi_ext.h>
+#include <rdma/providers/fi_peer.h>
 
 #include <ofi.h>
 #include <ofi_enosys.h>
