+++ prov/sm2/src/sm2.h
@@ -53,7 +53,7 @@
 #include <rdma/fi_tagged.h>
 #include <rdma/fi_trigger.h>
 #include <rdma/providers/fi_prov.h>
-#include <rdma/fi_ext.h>
+#include <rdma/providers/fi_peer.h>
 
 #include <ofi.h>
 #include <ofi_enosys.h>
