+++ src/fabric.c
@@ -50,7 +50,7 @@
 #include "ofi_prov.h"
 #include "ofi_perf.h"
 #include "ofi_hmem.h"
-#include "rdma/fi_ext.h"
+#include <rdma/fi_ext.h>
 
 #ifdef HAVE_LIBDL
 #include <dlfcn.h>
